+++ src/hand_tracker.py
@@ -3,7 +3,7 @@ import cv2
 import numpy as np
 import tensorflow as tf
 
-from non_maximum_suppression import non_max_suppression_fast
+from src.non_maximum_suppression import non_max_suppression_fast
 
 
 class HandTracker():
