+++ test/test_non_maximum_suppression.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 import numpy as np
 
-from non_maximum_suppression import non_max_suppression_fast
+from src.non_maximum_suppression import non_max_suppression_fast
 
 
 class Test(TestCase):
